Adds Yolanda Drew as TL to VILLA Camp 2 for Pre-Camp Days
Leslie Braff wrote a note on the top of the pay report that Yolanda worked 3 hours on Day -2 and 4 hours on Day -1. No clue what it is for, but I don't think it was tallied the first time around. This fix adds her to the list of TLs who are working Villa Camp 2.